has been updated webpack configuration;
