+++ package.json
@@ -20,6 +20,10 @@
     "babel-core": "^6.24.1",
     "babel-loader": "^6.4.1",
     "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.28.0",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "html-webpack-plugin": "^2.28.0",
+    "style-loader": "^0.16.1",
     "webpack": "^2.3.2"
   }
 }
