+++ src/js/common/common.js
