+++ src/news.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>Українські новини</title>
+    <link type="text/css" rel="stylesheet" href="style.css" media="all">
+</head>
+<body id="popup">
+<img id="title" src="../img/icon48.png">
+<h2>Українські новини</h2>
+<div id="toolbar">
+    <a href="options.html" title="Налаштування" target="_blank"><img src="../img/config.png"></a>
+    <a href="#" id="readall" title="Відмітити всі новини як прочитані"><img src="../img/delete.png"></a>
+</div>
+<div id="gradient"></div>
+<ul id="content">
+</ul>
+</body>
+</html>
