+++ src/options.html
@@ -3,7 +3,6 @@
 <head>
     <meta charset="UTF-8">
     <title>Українські новини - налаштування</title>
-    <script src="../src/js/common.js"></script>
     <link type="text/css" rel="stylesheet" href="options.css" media="all">
 </head>
 <body>
@@ -45,6 +44,5 @@
     </div>
 </div>
 <div class="copyright">© Права на матеріали належать сайтам-джерелам новин</div>
-<script src="../src/options.js"></script>
 </body>
 </html>
