+++ static/debug.htm
@@ -7,7 +7,7 @@
 <img id="title" src="img/icon48.png">
 <h2>Українські новини</h2>
 <div id="toolbar">
-    <a href="options.html" title="Налаштування" target="_blank"><img src="img/config.png"></a>
+    <a href="../src/options.html" title="Налаштування" target="_blank"><img src="img/config.png"></a>
     <a href="#" id="readall" title="Відмітити всі новини як прочитані"><img src="img/delete.png"></a>
 </div>
 <div id="gradient"></div>
