+++ static/options.css
@@ -1,6 +1,6 @@
 @font-face {
     font-family: titleFont;
-    src: url(resource/Arsenal-Regular.otf);
+    src: url(../../resource/Arsenal-Regular.otf);
 }
 
 body {
@@ -15,7 +15,7 @@ h2 {
     text-align: center;
     margin-bottom: 20px;
     color: #6f6f6f;
-    background: url(img/icon48.png) no-repeat left center;
+    background: url(../../img/icon48.png) no-repeat left center;
     background-size: 30px 30px;
     width: 310px;
 }
